azure.py
#!/usr/bin/env python3
"""
Complete Azure OpenAI Implementation
"""

# =============================================================================
# 1. ENVIRONMENT VARIABLES REQUIRED (.env file)
# =============================================================================

"""
# Authentication Variables (Required)
PING_FED_URL=your_ping_federation_oauth_endpoint
KGW_CLIENT_ID=your_client_id
KGW_CLIENT_SECRET=your_client_secret

# Azure OpenAI Configuration (Required)
KGW_ENDPOINT=your_azure_openai_endpoint
AOAI_API_VERSION=your_api_version
CHAT_MODEL_DEPLOYMENT_NAME=your_chat_model_deployment_name

# Optional Variables
GPT_O3_MINI_DEPLOYMENT_NAME=your_o3_mini_deployment_name
USE_O3_MINI=false
"""

# =============================================================================
# 2. API CALL IMPLEMENTATION
# =============================================================================

import os
import time
import requests
import json
from typing import Dict, Any, List, Optional

try:
    from dotenv import load_dotenv
    load_dotenv()
except ImportError:
    pass

def create_message_with_conversation(self, messages: List[Dict], max_tokens: int = 4000, temperature: float = 0.1) -> 'MockResponse':
    """Create a message with full conversation history (proper OpenAI format)"""
    # Get OAuth2 access token
    access_token = self.auth.get_access_token()
    
    # Construct Azure OpenAI endpoint URL
    url = f"{self.endpoint}/openai/deployments/{self.current_deployment}/chat/completions?api-version={self.api_version}"
    
    # Set headers with Bearer token authentication
    headers = {
        'Content-Type': 'application/json',
        'Authorization': f'Bearer {access_token}'
    }
    
    # Prepare OpenAI Chat Completions API payload
    payload = {
        'messages': messages,
        'max_completion_tokens': max_tokens,  # Use max_completion_tokens for newer models
        'temperature': temperature
    }
    
    # Retry logic with exponential backoff
    max_retries = 3
    for attempt in range(max_retries):
        try:
            # Make the HTTP POST request to Azure OpenAI
            response = requests.post(url, headers=headers, json=payload, timeout=120)
            
            if response.status_code == 200:
                # Success - parse response
                result = response.json()
                content = result['choices'][0]['message']['content']
                return MockResponse(content)
            elif response.status_code == 429:
                # Rate limit - wait and retry
                wait_time = (2 ** attempt) + 3
                print(f"   Rate limit hit, waiting {wait_time} seconds...")
                time.sleep(wait_time)
                continue
            else:
                # Other API errors
                error_msg = f"API Error: {response.status_code} - {response.text}"
                print(f"DEBUG: API Error - {error_msg}")
                return MockResponse(error_msg)
                
        except Exception as e:
            if attempt == max_retries - 1:
                # Final attempt failed
                error_msg = f"Request failed after {max_retries} attempts: {str(e)}"
                print(f"DEBUG: Request Error - {error_msg}")
                return MockResponse(error_msg)
            # Wait before retry
            wait_time = (2 ** attempt) + 2
            print(f"   Request error, retrying in {wait_time} seconds...")
            time.sleep(wait_time)
    
    return MockResponse(f"Failed to get response after {max_retries} attempts")

# =============================================================================
# 3. EXAMPLE USAGE
# =============================================================================

"""
# Example of how the variables are used in the client initialization:

def __init__(self):
    # Load authentication variables
    self.ping_fed_url = os.getenv('PING_FED_URL')           # OAuth2 endpoint
    self.kgw_client_id = os.getenv('KGW_CLIENT_ID')         # Client ID
    self.kgw_client_secret = os.getenv('KGW_CLIENT_SECRET') # Client Secret
    
    # Load Azure OpenAI configuration
    self.endpoint = os.getenv('KGW_ENDPOINT')                        # Azure endpoint
    self.api_version = os.getenv('AOAI_API_VERSION')                 # API version
    self.chat_deployment = os.getenv('CHAT_MODEL_DEPLOYMENT_NAME')   # Model deployment
    
    # Optional configuration
    self.o3_mini_deployment = os.getenv('GPT_O3_MINI_DEPLOYMENT_NAME')
    self.use_o3_mini = os.getenv('USE_O3_MINI', 'false').lower() == 'true'

# Example API call flow:
# 1. Get OAuth2 token using PING_FED_URL + KGW_CLIENT_ID + KGW_CLIENT_SECRET
# 2. Build URL: KGW_ENDPOINT + CHAT_MODEL_DEPLOYMENT_NAME + AOAI_API_VERSION
# 3. Make POST request with Bearer token
# 4. Parse response from choices[0].message.content
"""