requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.7
scikit-learn==1.3.2
numpy==1.24.3
python-multipart==0.0.6
pydantic==2.5.0
playwright==1.40.0
requests==2.31.0
